extract xbe::raw::Xbe to file
